+++ src/main/groovy/com/github/choonchernlim/calsync/exchange/ExchangeService.groovy
@@ -4,7 +4,6 @@ import com.github.choonchernlim.calsync.core.CalSyncEvent
 import com.github.choonchernlim.calsync.core.Mapper
 import com.github.choonchernlim.calsync.core.UserConfig
 import com.google.inject.Inject
-import microsoft.exchange.webservices.data.core.service.item.Appointment
 import org.joda.time.DateTime
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -34,24 +33,31 @@ class ExchangeService {
      * @param startDateTime Start datetime
      * @param endDateTime End datetime
      * @param includeCanceledEvents Whether to include canceled events or not
+     * @param includeEventBody Whether to include event body or not
      * @return Events if there's any, otherwise empty list
      */
-    List<CalSyncEvent> getEvents(DateTime startDateTime, DateTime endDateTime, Boolean includeCanceledEvents) {
+    List<CalSyncEvent> getEvents(
+            DateTime startDateTime,
+            DateTime endDateTime,
+            Boolean includeCanceledEvents,
+            Boolean includeEventBody) {
         assert startDateTime && endDateTime && startDateTime <= endDateTime
         assert includeCanceledEvents != null
+        assert includeEventBody != null
 
-        LOGGER.info("Retrieving events from ${startDateTime} to ${endDateTime}...")
+        LOGGER.info(
+                "Retrieving events from ${Mapper.humanReadableDateTime(startDateTime)} to ${Mapper.humanReadableDateTime(endDateTime)}...")
 
-        List<Appointment> exchangeEvents = exchangeClient.getEvents(startDateTime, endDateTime) ?: []
+        List<ExchangeEvent> exchangeEvents = exchangeClient.getEvents(startDateTime, endDateTime) ?: []
 
         LOGGER.info("\tTotal events found: ${exchangeEvents.size()}...")
 
         if (!includeCanceledEvents) {
-            exchangeEvents = exchangeEvents.findAll { !it.isCancelled }
+            exchangeEvents = exchangeEvents.findAll { !it.isCanceled }
             LOGGER.info("\tTotal events after excluding canceled events: ${exchangeEvents.size()}...")
         }
 
-        return exchangeEvents.collect { Mapper.toCalSyncEvent(it) }
+        return exchangeEvents.collect { Mapper.toCalSyncEvent(it, includeEventBody) }
     }
 }