+++ src/test/groovy/com/github/choonchernlim/calsync/core/MapperSpec.groovy
@@ -4,9 +4,6 @@ import com.google.api.client.util.DateTime
 import com.google.api.services.calendar.model.Event
 import com.google.api.services.calendar.model.EventDateTime
 import com.google.api.services.calendar.model.EventReminder
-import microsoft.exchange.webservices.data.core.ExchangeService
-import microsoft.exchange.webservices.data.core.enumeration.misc.ExchangeVersion
-import microsoft.exchange.webservices.data.core.service.item.Appointment
 import spock.lang.Specification
 import spock.lang.Unroll
 
@@ -92,34 +89,6 @@ class MapperSpec extends Specification {
         actual.reminderMinutesBeforeStart == 15
     }
 
-    def 'toCalSyncEvent - appointment - given valid value, should return calsync event'() {
-        given:
-        def startDateTime = org.joda.time.DateTime.now()
-        def endDateTime = startDateTime.plusDays(1)
-
-        def exchangeService = Mock(ExchangeService) {
-            getRequestedServerVersion() >> ExchangeVersion.Exchange2010_SP2
-        }
-        def appointment = new Appointment(exchangeService)
-
-        appointment.start = startDateTime.toDate()
-        appointment.end = endDateTime.toDate()
-        appointment.subject = 'summary'
-        appointment.location = 'location'
-        appointment.reminderMinutesBeforeStart = 15
-
-        when:
-        def actual = Mapper.toCalSyncEvent(appointment)
-
-        then:
-        actual.googleEventId == null
-        actual.startDateTime == startDateTime
-        actual.endDateTime == endDateTime
-        actual.subject == 'summary'
-        actual.location == 'location'
-        actual.reminderMinutesBeforeStart == 15
-    }
-
     def 'toGoogleEvent - given valid value, should return google event'() {
         given:
         def startDateTime = org.joda.time.DateTime.now()
@@ -149,4 +118,61 @@ class MapperSpec extends Specification {
         reminders[0].getMinutes() == 15
         reminders[0].getMethod() == 'popup'
     }
+
+    @Unroll
+    def 'toPlainText - given #label value, should be null'() {
+        when:
+        def actual = Mapper.toPlainText(input)
+
+        then:
+        actual == null
+
+        where:
+        label   | input
+        'null'  | null
+        'blank' | ' '
+    }
+
+    def 'toPlainText - given valid HTML, should return plain text'() {
+        given:
+        def input = '<html><head><meta http-equiv="Content-Type" content="text/html; charset=utf-8"><meta name="Generator" content="Microsoft Exchange Server"><!-- converted from rtf --><style><!-- .EmailQuote { margin-left: 1pt; padding-left: 4pt; border-left: #800000 2px solid; } --></style></head><body><font face="Calibri" size="2"><span style="font-size:11pt;"><div>123 / 456</div><div>&nbsp;</div><div>12/9/2016 Added room location – limc</div><div>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; Room Reservation # 999</div><div>&nbsp;</div><div>&nbsp;</div></span></font></body></html>'
+
+        when:
+        def actual = Mapper.toPlainText(input)
+
+        then:
+        actual == '''123 / 456 
+
+12/9/2016 Added room location – limc
+
+        Room Reservation # 999'''
+    }
+
+    def 'toPlainText - given valid HTML but no content, should return null'() {
+        given:
+        def input = '<html><div></div></html>'
+
+        when:
+        def actual = Mapper.toPlainText(input)
+
+        then:
+        actual == null
+    }
+
+    def 'humanReadableDateTime - given null value, should throw exception'() {
+        when:
+        Mapper.humanReadableDateTime(null)
+
+        then:
+        thrown AssertionError
+    }
+
+    def 'humanReadableDateTime - given valid value, should return formatted datetime'() {
+        when:
+        def actual = Mapper.humanReadableDateTime(new org.joda.time.DateTime(2016, 12, 12, 9, 10, 11, 12))
+
+        then:
+        actual == 'Dec 12 @ 09:10 AM'
+    }
+
 }