+++ src/test/groovy/com/github/choonchernlim/calsync/exchange/ExchangeServiceSpec.groovy
@@ -1,8 +1,6 @@
 package com.github.choonchernlim.calsync.exchange
 
 import com.github.choonchernlim.calsync.core.UserConfig
-import microsoft.exchange.webservices.data.core.enumeration.misc.ExchangeVersion
-import microsoft.exchange.webservices.data.core.service.item.Appointment
 import org.joda.time.DateTime
 import spock.lang.Specification
 import spock.lang.Unroll
@@ -37,7 +35,7 @@ class ExchangeServiceSpec extends Specification {
     @Unroll
     def 'getEvents - given #label, should throw exception'() {
         when:
-        service.getEvents(startDateTime, endDateTime, true)
+        service.getEvents(startDateTime, endDateTime, true, true)
 
         then:
         0 * _
@@ -56,7 +54,20 @@ class ExchangeServiceSpec extends Specification {
         def endDateTime = startDateTime.plusDays(1)
 
         when:
-        service.getEvents(startDateTime, endDateTime, null)
+        service.getEvents(startDateTime, endDateTime, null, true)
+
+        then:
+        0 * _
+        thrown AssertionError
+    }
+
+    def 'getEvents - given includeEventBody == null, should throw exception'() {
+        given:
+        def startDateTime = DateTime.now()
+        def endDateTime = startDateTime.plusDays(1)
+
+        when:
+        service.getEvents(startDateTime, endDateTime, true, null)
 
         then:
         0 * _
@@ -70,7 +81,7 @@ class ExchangeServiceSpec extends Specification {
         def endDateTime = startDateTime.plusDays(1)
 
         when:
-        def events = service.getEvents(startDateTime, endDateTime, true)
+        def events = service.getEvents(startDateTime, endDateTime, true, true)
 
         then:
         1 * exchangeClient.getEvents(startDateTime, endDateTime) >> null
@@ -79,50 +90,86 @@ class ExchangeServiceSpec extends Specification {
         events.isEmpty()
     }
 
-    def 'getEvents - given events found, should return events'() {
+    def 'getEvents - given includeCanceledEvent == true and includeEventBody == true, should return all events with body'() {
         given:
         def startDateTime = DateTime.now()
         def endDateTime = startDateTime.plusDays(1)
 
         when:
-        def events = service.getEvents(startDateTime, endDateTime, true)
+        def events = service.getEvents(startDateTime, endDateTime, true, true)
 
         then:
         1 * exchangeClient.getEvents(startDateTime, endDateTime) >> [
-                createAppointment(startDateTime, endDateTime, 'summary1', 'location1', 15, true),
-                createAppointment(startDateTime, endDateTime, 'summary2', 'location2', 15, false)
+                createExchangeEvent(startDateTime, endDateTime, 'summary1', 'location1', 15, true, 'body1'),
+                createExchangeEvent(startDateTime, endDateTime, 'summary2', 'location2', 15, false, 'body2')
         ]
         0 * _
 
         events.size() == 2
         events*.subject == ['summary1', 'summary2']
+        events*.body == ['body1', 'body2']
     }
 
-    private Appointment createAppointment(
+    def 'getEvents - given includeCanceledEvent == false and includeEventBody == true, should return all non-canceled events with body'() {
+        given:
+        def startDateTime = DateTime.now()
+        def endDateTime = startDateTime.plusDays(1)
+
+        when:
+        def events = service.getEvents(startDateTime, endDateTime, false, true)
+
+        then:
+        1 * exchangeClient.getEvents(startDateTime, endDateTime) >> [
+                createExchangeEvent(startDateTime, endDateTime, 'summary1', 'location1', 15, true, 'body1'),
+                createExchangeEvent(startDateTime, endDateTime, 'summary2', 'location2', 15, false, 'body2')
+        ]
+        0 * _
+
+        events.size() == 1
+        events*.subject == ['summary2']
+        events*.body == ['body2']
+    }
+
+    def 'getEvents - given includeCanceledEvent == true and includeEventBody == false, should return all events with no body'() {
+        given:
+        def startDateTime = DateTime.now()
+        def endDateTime = startDateTime.plusDays(1)
+
+        when:
+        def events = service.getEvents(startDateTime, endDateTime, true, false)
+
+        then:
+        1 * exchangeClient.getEvents(startDateTime, endDateTime) >> [
+                createExchangeEvent(startDateTime, endDateTime, 'summary1', 'location1', 15, true, 'body1'),
+                createExchangeEvent(startDateTime, endDateTime, 'summary2', 'location2', 15, false, 'body2')
+        ]
+        0 * _
+
+        events.size() == 2
+        events*.subject == ['summary1', 'summary2']
+        events*.body == [null, null]
+    }
+
+    private static ExchangeEvent createExchangeEvent(
             DateTime startDateTime,
             DateTime endDateTime,
             String subject,
             String location,
             Integer reminderMinutesBeforeStart,
-            Boolean isCancelled) {
+            Boolean isCancelled,
+            String body) {
 
-        def exchangeService = Mock(microsoft.exchange.webservices.data.core.ExchangeService) {
-            getRequestedServerVersion() >> ExchangeVersion.Exchange2010_SP2
-        }
 
-        def appointment = new Appointment(exchangeService)
+        return new ExchangeEvent(
+                startDateTime: startDateTime,
+                endDateTime: endDateTime,
+                subject: subject,
+                location: location,
+                reminderMinutesBeforeStart: reminderMinutesBeforeStart,
+                isCanceled: isCancelled,
+                body: body
 
-        // TODO  You must load or assign this property before you can read its value.
-        // appointment.load(new PropertySet(BasePropertySet.FirstClassProperties))
-
-        appointment.start = startDateTime.toDate()
-        appointment.end = endDateTime.toDate()
-        appointment.subject = subject
-        appointment.location = location
-        appointment.reminderMinutesBeforeStart = reminderMinutesBeforeStart
-        appointment.isCancelled == isCancelled
-
-        return appointment
+        )
     }