+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.2.0 - 2016-12-13
+
+* Two new configuration options:-
+    * `include.canceled.events` = Whether to include canceled events or not.
+    * `include.event.body` = Whether to include event body or not.
+
+* Better console log messages.
+
 ## 0.1.0 - 2016-12-11
 
 * Initial.